JpaGoalDao: Changed return type and exceptions according to GoalDao interface and changed logic of updateGoal method

+++ src/main/java/cz/muni/fi/pa165/mamatoad/soccerrecords/goal/JpaGoalDao.java
@@ -7,6 +7,7 @@ package cz.muni.fi.pa165.mamatoad.soccerrecords.goal;
 import cz.muni.fi.pa165.mamatoad.soccerrecords.match.Match;
 import cz.muni.fi.pa165.mamatoad.soccerrecords.player.Player;
 import cz.muni.fi.pa165.mamatoad.soccerrecords.team.Team;
+import cz.muni.fi.pa165.mamatoad.soccerrecords.util.exception.IllegalEntityException;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
@@ -14,7 +15,7 @@ import javax.persistence.TypedQuery;
 
 /**
  *
- * @author Adriana
+ * @author Adriana Smijakova
  */
 public class JpaGoalDao implements GoalDao {
 
@@ -25,40 +26,68 @@ public class JpaGoalDao implements GoalDao {
     }
 
     @Override
-    public Goal createGoal(Goal goal) throws IllegalArgumentException {
+    public void createGoal(Goal goal) throws IllegalArgumentException, IllegalEntityException {
         if (goal == null) {
             throw new IllegalArgumentException("Parameter goal is null");
         }
+        
+        StringBuilder builder = new StringBuilder("These parameters are null: ");
+        
+        boolean throwEx = validateGoal(goal, builder);
+        if(goal.getId() != null){
+            builder.append("\n Id should be null.");
+        }
+        if(throwEx){
+            throw new IllegalEntityException(builder.toString());
+        }
 
         EntityManager em = factory.createEntityManager();
+        
         em.getTransaction().begin();
         em.persist(goal);
         em.getTransaction().commit();
-
-        return goal;
     }
 
     @Override
-    public Goal updateGoal(Goal goal) throws IllegalArgumentException {
+    public void updateGoal(Goal goal) throws IllegalArgumentException, IllegalEntityException {
         if (goal == null) {
             throw new IllegalArgumentException("Parameter goal is null");
         }
+        
+        StringBuilder builder = new StringBuilder("These parameters are null: ");
+        
+        boolean throwEx = validateGoal(goal, builder);
+        if(goal.getId() == null){
+            builder.append("id ");
+        }
+        if(throwEx){
+            throw new IllegalEntityException(builder.toString());
+        }
 
         EntityManager em = factory.createEntityManager();
-        em.getTransaction().begin();
-        em.persist(goal);
-        em.getTransaction().commit();
-
-        return goal;
+        
+        if(em.find(Goal.class, goal) == null){
+            throw new IllegalEntityException("Goal is not in database");
+        }
+        
+        em.merge(goal);
     }
 
     @Override
-    public void deleteGoal(Goal goal) throws IllegalArgumentException {
+    public void deleteGoal(Goal goal) throws IllegalArgumentException, IllegalEntityException {
         if (goal == null) {
             throw new IllegalArgumentException("Parameter goal is null");
         }
+        if(goal.getId() == null){
+            throw new IllegalEntityException("Goal id is null");
+        }
 
         EntityManager em = factory.createEntityManager();
+        
+        if(em.find(Goal.class, goal) == null){
+            throw new IllegalEntityException("Goal is not in database");
+        }
+        
         em.getTransaction().begin();
         em.remove(goal);
         em.getTransaction().commit();
@@ -77,9 +106,12 @@ public class JpaGoalDao implements GoalDao {
     }
 
     @Override
-    public List<Goal> retrieveGoalsByMatch(Match match) throws IllegalArgumentException {
+    public List<Goal> retrieveGoalsByMatch(Match match) throws IllegalArgumentException, IllegalEntityException {
         if (match == null) {
             throw new IllegalArgumentException("Parameter goal is null");
+        }
+        if (match.getId() == null){
+            throw new IllegalEntityException("Id of match is null");
         }
 
         EntityManager em = factory.createEntityManager();
@@ -90,12 +122,18 @@ public class JpaGoalDao implements GoalDao {
     }
 
     @Override
-    public List<Goal> retrieveGoalsByMatchAndTeam(Match match, Team team) throws IllegalArgumentException {
+    public List<Goal> retrieveGoalsByMatchAndTeam(Match match, Team team) throws IllegalArgumentException, IllegalEntityException {
         if (match == null) {
             throw new IllegalArgumentException("Parameter goal is null");
         }
         if (team == null) {
             throw new IllegalArgumentException("Parameter team is null");
+        }
+        if(match.getId() == null){
+            throw new IllegalEntityException("Match id is null");
+        }
+        if(team.getId() == null){
+            throw new IllegalEntityException("Team id is null");
         }
 
         EntityManager em = factory.createEntityManager();
@@ -108,9 +146,12 @@ public class JpaGoalDao implements GoalDao {
     }
 
     @Override
-    public List<Goal> retrieveGoalsByPlayer(Player player) throws IllegalArgumentException {
+    public List<Goal> retrieveGoalsByPlayer(Player player) throws IllegalArgumentException, IllegalEntityException {
         if (player == null) {
             throw new IllegalArgumentException("Parameter player is null");
+        }
+        if(player.getId() == null){
+            throw new IllegalEntityException("Player id is null");
         }
 
         EntityManager em = factory.createEntityManager();
@@ -121,12 +162,18 @@ public class JpaGoalDao implements GoalDao {
     }
 
     @Override
-    public List<Goal> retrieveGoalsByMatchAndPlayer(Match match, Player player) throws IllegalArgumentException {
+    public List<Goal> retrieveGoalsByMatchAndPlayer(Match match, Player player) throws IllegalArgumentException, IllegalEntityException {
        if (match == null) {
             throw new IllegalArgumentException("Parameter goal is null");
         }
         if (player == null) {
             throw new IllegalArgumentException("Parameter player is null");
+        }
+        if(match.getId() == null){
+            throw new IllegalEntityException("Match id is null");
+        }
+        if(player.getId() == null){
+            throw new IllegalEntityException("Player id is null");
         }
 
         EntityManager em = factory.createEntityManager();
@@ -136,5 +183,28 @@ public class JpaGoalDao implements GoalDao {
         goals.setParameter("player", player);
 
         return goals.getResultList();
+    }
+    
+    static private boolean validateGoal(Goal goal, StringBuilder builder){
+        boolean throwEx = false;
+        
+        if(goal.getMatch() == null) {
+            builder.append("match ");
+            throwEx = true;
+        }
+        if(goal.getPlayer() == null){
+            builder.append("player ");
+            throwEx = true;
+        }
+        if(goal.getTeam() == null){
+            builder.append("team ");
+            throwEx = true;
+        }
+        if(goal.getShootingTime() == null){
+            builder.append("shootingTime ");
+            throwEx = true;
+        }
+        
+        return throwEx;
     }
 }
